Update README profile info and add badges
This commit expands the README profile introduction and adds different sections such as 'About me', 'Skills', 'Current projects', and 'To contact me'. The update aims to provide a more comprehensive and visually appealing overview of the profile. It also replaces the standard salutation.

Addition of shields.io badges prominently displays the tools, languages and projects in a more streamlined and colorful way, hence enhancing the profile's visual appeal and clarity.